Refactor - export convert_to_download_url
